requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
django-bootstrap5==24.3
Faker==33.3.1
django-extensions==3.2.3
